### 2025-07-31 #254 @HzaCode
ChemInformant – Python client for PubChem
Submitting Author: **Zhiang He (@HzaCode)**
All current maintainers: **Zhiang He (@HzaCode)**
Package Name: **ChemInformant**
One-Line Description of Package: **A robust, high-throughput Python client for retrieving chemical information from the PubChem API; it returns analysis-ready Pandas/SQL outputs, handles caching, rate-limiting and retries, and includes convenient CLI tools.**
Repository Link: **https://github.com/HzaCode/ChemInformant**
Version submitted: **2.4.0 (released 2025-07-30)**
Reviewer 1:
Reviewer 2:
Archive: 10.5281/zenodo.16623785
JOSS DOI: https://doi.org/10.21105/joss.08341
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

ChemInformant is a workflow-centric Python client for the PubChem REST API.
It transforms raw API responses into clean Pandas DataFrames or direct SQL tables, supports mixed identifiers (CIDs, names, SMILES), provides persistent on-disk caching, automatic rate-limiting and retries, and exposes both an object-oriented API and terminal-ready CLI tools (`chemfetch`, `chemdraw`).
The library leverages Pydantic v2 for runtime data validation and integrates seamlessly into machine-learning or QSAR pipelines that require large-scale descriptor retrieval.

If your package is associated with an existing community please check below:

- [ ] Astropy: My package adheres to Astropy community standards

### Scope Explanation

* **Target audience & applications:** Computational chemists, cheminformatics researchers, drug-discovery scientists and data-science practitioners who require reliable, programmatic access to PubChem for large-scale descriptor retrieval and downstream ML/QSAR workflows.
* **Comparison to similar packages:** Existing clients such as *PubChemPy* or *ChemSpiPy* lack built-in caching, mixed-identifier handling and robust fault-tolerance; in benchmarks, ChemInformant achieves up to a 48× speed-up versus PubChemPy through batched queries and cache reuse.
* **Pre-submission enquiry:** None.

- [x] uses an [OSI approved license][OsiApprovedLicense] (MIT).
- [x] has a test suite (`pytest`) located in `tests/`.
- [x] has continuous integration setup via GitHub Actions.

* [ ] **Do you wish to automatically submit to the \[Journal of Open Source Software]\[JournalOfOpenSourceSoftware]?**

  *ChemInformant* has **been accepted by JOSS**
  ([https://github.com/openjournals/joss-reviews/issues/8341](https://github.com/openjournals/joss-reviews/issues/8341)).

  I would be grateful if pyOpenSci could adopt the **JOSS → pyOpenSci fast-track** pathway so that a second full technical review is not required.

  Please let me know if any additional information or metadata would be helpful—thank you very much for your time and consideration.

  <summary>JOSS Checks (leave blank since box above is unchecked)</summary>

- [ ] [Last but not least please fill out our pre-review survey](https://forms.gle/F9mou7S3jhe8DMJ16). This helps us track submission and improve our peer review process. We will also ask our reviewers and editors to fill this out.
