### 2025-08-05 #254 @HzaCode
Thanks for letting me know! I'm really looking forward to the process and hope to contribute to the pyOpenSci community. 💪
