### 2025-08-13 #254 @HzaCode
Hi @eliotwrobson,

Just a quick update — my JOSS paper on this project has now been accepted. I’m looking forward to the pyOpenSci process as well — are there any next steps or timelines you could share? Thanks!
