### 2025-08-20 #254 @eliotwrobson
@HzaCode So we don't have a process in place yet for automatic JOSS -> pyOS acceptance (the process you were referring to originally was the reverse), but we think that having this would be a good idea, so we are in talks to add some process behind this and get your package approved soon 👍🏽
