### 2025-08-20 #254 @HzaCode
Hi @eliotwrobson,

Thank you so much for the positive update! I'm thrilled to hear that you and the team are considering establishing a JOSS -> pyOpenSci fast-track pathway.

I completely agree that this is a fantastic idea. It would not only save a great deal of time for authors who have already passed the JOSS review but also allow pyOpenSci's resources to focus on its community-specific standards, making the entire ecosystem more efficient.

My project, ChemInformant, would be very happy to serve as the **first** case for this new process. Please let me know if there's anything I can do to assist as you establish it.

Thanks again for your foresight and your dedication to the community.
