### 2025-08-22 32cf3ac
Update coverage badge [skip ci] 160757180a06cd282fc09fa4f1825084b124afaa
