### 2025-08-24 c993506
Update coverage badge [skip ci] 0589ac9dde66f3a82fb8486167a2530bc4f4976d
