### 2025-09-10 30b6d56
v2.4.2: Code quality improvements and ruff fixes
