+++ src/ChemInformant/cli.py
@@ -7,11 +7,11 @@ import sys
 
 # Use relative import because cli.py is inside the ChemInformant package
 from .cheminfo_api import (
-    get_properties,
     draw_compound,
+    get_properties,
 )
-from .models import NotFoundError, AmbiguousIdentifierError
 from .constants import ALL_PROPS
+from .models import AmbiguousIdentifierError, NotFoundError
 
 # --- NEUE ERGÄNZUNG: Importieren der SQL-Helferfunktion ---
 from .sql import df_to_sql
@@ -22,41 +22,41 @@ from .sql import df_to_sql
 def main_fetch():
     """
     Main entry point for the chemfetch command-line tool.
-    
+
     Provides a command-line interface for retrieving chemical data from PubChem
     with support for multiple output formats including CSV, JSON, and SQL databases.
     Supports all ChemInformant features including batch processing, property selection,
     and the new snake_case standardized format.
-    
+
     Command format:
         chemfetch [identifiers...] [options]
-        
+
     Key features:
         - Multiple output formats: CSV, JSON, SQL database
         - Property selection: --props for specific properties
         - Batch modes: --all-properties, --include-3d
         - Direct SQL export: --format sql -o database.db
         - Error handling: Shows failed lookups with status information
-        
+
     Examples:
         # Basic usage - core properties to CSV
         chemfetch aspirin caffeine
-        
+
         # Specific properties to JSON
         chemfetch aspirin --props molecular_weight,xlogp --format json
-        
+
         # All properties to SQL database
         chemfetch aspirin caffeine --all-properties --format sql -o chemicals.db
-        
+
         # Include 3D descriptors
         chemfetch aspirin --include-3d --format csv -o results.csv
     """
-    available_props = sorted(ALL_PROPS)
+    sorted(ALL_PROPS)
 
     parser = argparse.ArgumentParser(
         prog="chemfetch",
         description="ChemInformant: A command-line tool to fetch chemical data from PubChem.",
-        epilog=f"Example: chemfetch aspirin caffeine --props cas,molecular_weight --format sql -o results.db",
+        epilog="Example: chemfetch aspirin caffeine --props cas,molecular_weight --format sql -o results.db",
     )
 
     parser.add_argument(
@@ -152,27 +152,27 @@ def main_fetch():
 def main_draw():
     """
     Main entry point for the chemdraw command-line tool.
-    
+
     Provides a simple command-line interface for visualizing 2D chemical structures
     directly from the terminal. Fetches structure images from PubChem and displays
     them using matplotlib.
-    
+
     Command format:
         chemdraw [identifier]
-        
+
     Features:
         - Accepts any ChemInformant identifier (name, CID, SMILES)
         - Displays structure with compound name as title
         - Requires matplotlib and PIL (install with: pip install ChemInformant[plot])
         - High-quality PNG images from PubChem
-        
+
     Examples:
         # Visualize by name
         chemdraw aspirin
-        
+
         # Visualize by CID
         chemdraw 2244
-        
+
         # Visualize by SMILES
         chemdraw "CC(=O)OC1=CC=CC=C1C(=O)O"
     """
