+++ docs/source/conf.py
@@ -32,9 +32,9 @@ html_theme       = "sphinx_rtd_theme"
 html_static_path = ["_static"]
 
 html_theme_options = {
-    "collapse_navigation": True,   
-    "navigation_depth"  : 2,       
-    "style_external_links": True,  
+    "collapse_navigation": True,
+    "navigation_depth"  : 2,
+    "style_external_links": True,
 }
 
 # -- Intersphinx --------------------------------------------------------------
