### 2025-09-19 #254 @eliotwrobson
@HzaCode thanks for being patient here. We just approved our JOSS -> pyOS process the other day, so I will be able to conduct the review of this package this weekend 👍🏽 Link is here https://www.pyopensci.org/software-peer-review/our-process/policies.html#expedited-review-process
