### 2025-09-19 #254 @HzaCode
I'm grateful for the pyOpenSci team's trust and efficiency, and I'm happy I could contribute to the community in a way that goes beyond the code.Helping to drive **a more efficient and collaborative process** for the open-source ecosystem brings a sense of accomplishment that rivals the joy of the project itself.
