### 2025-09-22 #254 @eliotwrobson
Since the process for the accelerated acceptance is brand new, I'm happy to serve as editor here. The requirements for just an editor going over the review checklist are not very detailed yet (it's not even clear that I should fill out a survey before doing this), but I think it's sufficient to go through the reviewer checklist and then touch base with others to see what type of wrap-up is needed.

@HzaCode thanks again for your submission! The package looks like it's in great shape, just a few comments to resolve. In the meantime, I'll check with other editors to make sure the acceptance process goes smoothly 👍🏽

- [x] As the reviewer, I confirm that there are no conflicts of interest for me to review this work (If you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).

- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience in the README file.
- [x] **Short quickstart tutorials** demonstrating significant functionality that successfully runs locally.
- [x] **Metadata** including author(s), author e-mail(s), a URL, and any other relevant metadata, for example, in a `pyproject.toml` file or elsewhere.

*NOTE: If the README has many more badges, you might want to consider using a table for badges: [see this example](https://github.com/ropensci/drake). Such a table should be wider than high. A badge for pyOpenSci peer review will be provided when the package is accepted.*

The package structure should follow the general community best practices. In general, please consider whether:

---

- For the CI badges, be sure to add a badge for the docs building in addition to the functional code tests passing.
- You have a few different vignettes that cover different use cases, it would be nice to have a few of these linked in the README directly.
- The documentation site seems to have inconsistent syntax highlighting, and this would make the code much easier to read.
- It looks like your code does not have a lint workflow for automated style checking in the library code. Adding this to your CI would make your code much easier to maintain, along with being a requirement for acceptance. We suggest ruff as a good tool for this.
- Similarly, your code includes type annotations, but no mypy workflow in CI or py.typed file to export the annotations. Having both of these would improve your code quality and make it easier for client code to use your package.
- It looks like your package supports versions >= 3.8, but this has been EOL for a while. You should drop support for old versions.
- For your benchmarks, I think it would be cleaner to add them to your test suite and use pytest-benchmark. Then it will be easier for users to download and run, and the output can be part of the standard test reports.
