### 2025-09-22 5a826e3
Add docs badge and more README links
