### 2025-09-22 ba2719a
Fix syntax highlighting in docs
