### 2025-09-22 958b4e8
Add ruff and mypy to CI
