### 2025-09-22 fc6bfef
Drop Python 3.8 support
