### 2025-09-22 1673edf
Switch to pytest-benchmark
