+++ tests/test_benchmarks.py
@@ -1,0 +1,227 @@
+"""
+Benchmark tests for ChemInformant performance using pytest-benchmark.
+
+This module replaces the standalone benchmark.py script with integrated
+pytest benchmarks that can be run as part of the standard test suite.
+"""
+
+import statistics
+import time
+
+import pytest
+
+import ChemInformant as ci
+
+# Sample drug names for benchmarking
+SAMPLE_DRUGS = [
+    "aspirin", "caffeine", "acetaminophen", "ibuprofen", "naproxen",
+    "diclofenac", "celecoxib", "meloxicam", "indomethacin", "piroxicam",
+    "amoxicillin", "azithromycin", "ciprofloxacin", "doxycycline", "penicillin",
+    "atorvastatin", "simvastatin", "rosuvastatin", "pravastatin", "lovastatin",
+    "metformin", "insulin", "glipizide", "glyburide", "pioglitazone",
+    "lisinopril", "enalapril", "losartan", "valsartan", "amlodipine",
+    "metoprolol", "atenolol", "propranolol", "carvedilol", "bisoprolol",
+    "warfarin", "heparin", "clopidogrel", "aspirin", "rivaroxaban",
+    "omeprazole", "lansoprazole", "pantoprazole", "esomeprazole", "ranitidine",
+    "fluoxetine", "sertraline", "paroxetine", "citalopram", "escitalopram",
+]
+
+BENCHMARK_PROPERTIES = [
+    "molecular_weight",
+    "xlogp",
+    "cas",
+    "iupac_name",
+    "canonical_smiles",
+    "molecular_formula"
+]
+
+
+class TestPerformanceBenchmarks:
+    """Performance benchmarks for ChemInformant core functionality."""
+
+    @pytest.mark.benchmark(group="single_compound")
+    def test_single_compound_lookup(self, benchmark):
+        """Benchmark single compound property lookup."""
+        def single_lookup():
+            return ci.get_properties(["aspirin"], ["molecular_weight", "xlogp"])
+
+        result = benchmark(single_lookup)
+        assert len(result) == 1
+        assert result.iloc[0]["status"] == "OK"
+
+    @pytest.mark.benchmark(group="batch_processing")
+    def test_small_batch_processing(self, benchmark):
+        """Benchmark small batch processing (10 compounds)."""
+        compounds = SAMPLE_DRUGS[:10]
+
+        def batch_lookup():
+            return ci.get_properties(compounds, BENCHMARK_PROPERTIES[:3])
+
+        result = benchmark(batch_lookup)
+        assert len(result) <= len(compounds)  # Some compounds might fail
+
+    @pytest.mark.benchmark(group="batch_processing")
+    def test_medium_batch_processing(self, benchmark):
+        """Benchmark medium batch processing (25 compounds)."""
+        compounds = SAMPLE_DRUGS[:25]
+
+        def batch_lookup():
+            return ci.get_properties(compounds, BENCHMARK_PROPERTIES[:4])
+
+        result = benchmark(batch_lookup)
+        assert len(result) <= len(compounds)
+
+    @pytest.mark.benchmark(group="batch_processing")
+    def test_large_batch_processing(self, benchmark):
+        """Benchmark large batch processing (50 compounds)."""
+        compounds = SAMPLE_DRUGS
+
+        def batch_lookup():
+            return ci.get_properties(compounds, BENCHMARK_PROPERTIES)
+
+        result = benchmark(batch_lookup)
+        assert len(result) <= len(compounds)
+
+    @pytest.mark.benchmark(group="caching")
+    def test_cache_performance(self, benchmark):
+        """Benchmark cache hit performance."""
+        # First, populate the cache
[…]
