### 2025-09-22 1b5cad1
Apply ruff formatting
