+++ src/ChemInformant/api_helpers.py
@@ -5,6 +5,7 @@ and provides low-level functions that wrap specific PubChem PUG-REST endpoints.
 These functions are not intended for direct use by end-users, but are consumed
 by the main API interface.
 """
+
 from __future__ import annotations
 
 import random
@@ -18,8 +19,8 @@ import requests_cache
 
 # --- Module Constants ---
 PUBCHEM_API_BASE = "https://pubchem.ncbi.nlm.nih.gov/rest/pug"
-PUG_VIEW_BASE    = "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data"
-REQUEST_TIMEOUT  = 15
+PUG_VIEW_BASE = "https://pubchem.ncbi.nlm.nih.gov/rest/pug_view/data"
+REQUEST_TIMEOUT = 15
 
 MAX_RETRIES, INITIAL_BACKOFF, MAX_BACKOFF = 5, 1, 16
 REQUEST_RATE_LIMIT = 5  # Requests per second
@@ -31,6 +32,7 @@ _session: requests_cache.CachedSession | None = None
 
 
 # --- Session & Caching ---
+
 
 def setup_cache(
     cache_name: str = "pubchem_cache",
@@ -56,12 +58,17 @@ def setup_cache(
     """
     global _session
     _session = requests_cache.CachedSession(
-        cache_name     = cache_name,
-        backend        = backend,
-        expire_after   = expire_after,
-        allowable_codes= [200, 404, 503],  # Cache "not found" and "server busy" responses
+        cache_name=cache_name,
+        backend=backend,
+        expire_after=expire_after,
+        allowable_codes=[
+            200,
+            404,
+            503,
+        ],  # Cache "not found" and "server busy" responses
         **kw,
     )
+
 
 def get_session() -> requests_cache.CachedSession:
     """
@@ -80,6 +87,7 @@ def get_session() -> requests_cache.CachedSession:
 
 # --- Core Fetching Logic ---
 
+
 def _execute_fetch(url: str) -> requests.Response:
     """
     Executes a single GET request using the global session.
@@ -91,7 +99,10 @@ def _execute_fetch(url: str) -> requests.Response:
     """
     return get_session().get(url, timeout=REQUEST_TIMEOUT)
 
-def _fetch_with_ratelimit_and_retry(url: str) -> dict[str, Any] | list[Any] | str | None:
+
+def _fetch_with_ratelimit_and_retry(
+    url: str,
+) -> dict[str, Any] | list[Any] | str | None:
     """
     Performs a GET request with rate-limiting and exponential backoff retry logic.
 
@@ -135,22 +146,34 @@ def _fetch_with_ratelimit_and_retry(url: str) -> dict[str, Any] | list[Any] | st
                 return None  # Resource not found is a valid, final state.
 
             if resp.status_code == 503:
-                print(f"[ChemInformant] 503 Server Busy -> retry in {backoff:.1f}s", file=sys.stderr)
+                print(
+                    f"[ChemInformant] 503 Server Busy -> retry in {backoff:.1f}s",
+                    file=sys.stderr,
+                )
             else:
                 resp.raise_for_status()  # Trigger for other 4xx/5xx errors
 
         except requests.exceptions.RequestException as e:
-            print(f"[ChemInformant] Network error {e} -> retry in {backoff:.1f}s", file=sys.stderr)
+            print(
+                f"[ChemInformant] Network error {e} -> retry in {backoff:.1f}s",
+                file=sys.stderr,
+            )
 
         time.sleep(backoff)
-        backoff = min(MAX_BACKOFF, backoff * 2) + random.uniform(0, 1) # Exponential backoff with jitter
[…]
