+++ src/ChemInformant/cheminfo_api.py
@@ -2,6 +2,7 @@
 This module contains the high-level, user-facing API for ChemInformant.
 It now operates exclusively on a snake_case standard for all property names.
 """
+
 from __future__ import annotations
 
 import logging
@@ -25,8 +26,10 @@ from .models import AmbiguousIdentifierError, Compound, NotFoundError
 _SMILES_TOKENS = re.compile(r"[=#\[\]\(\)]|\d|Br|Cl|Si", re.I)
 _SPECIAL_PROPS = {"cas", "synonyms"}
 
+
 def _looks_like_smiles(s: str) -> bool:
     return bool(_SMILES_TOKENS.search(s))
+
 
 def _resolve_to_single_cid(identifier: str | int) -> int:
     # This function remains robust as is.
@@ -49,6 +52,7 @@ def _resolve_to_single_cid(identifier: str | int) -> int:
                 return cids[0]
             raise AmbiguousIdentifierError(identifier, cids)
     raise NotFoundError(identifier)
+
 
 def get_properties(
     identifiers: int | str | list[int] | list[str],
@@ -144,7 +148,9 @@ def get_properties(
     """
     # --- Parameter validation ---
     if all_properties and (properties is not None or include_3d):
-        raise ValueError("`all_properties=True` is mutually exclusive with `properties` and `include_3d`.")
+        raise ValueError(
+            "`all_properties=True` is mutually exclusive with `properties` and `include_3d`."
+        )
     if properties is not None and include_3d:
         logging.warning("When `properties` is specified, `include_3d` is ignored.")
 
@@ -156,7 +162,7 @@ def get_properties(
         original_props_order = resolved_props
     elif properties is not None:
         if isinstance(properties, str):
-            properties = [p.strip() for p in properties.split(',')]
+            properties = [p.strip() for p in properties.split(",")]
         original_props_order = list(properties)
         resolved_props = []
         unsupported_props = []
@@ -169,7 +175,7 @@ def get_properties(
 
         if unsupported_props:
             raise ValueError(f"Unsupported properties: {unsupported_props}")
-    else: # Default behavior when properties is None
+    else:  # Default behavior when properties is None
         resolved_props = CORE_PROPS[:]
         if include_3d:
             resolved_props.extend(THREED_PROPS)
@@ -190,7 +196,9 @@ def get_properties(
             meta[ident] = {"status": "OK", "cid": str(cid)}
         except Exception as exc:
             meta[ident] = {"status": type(exc).__name__, "cid": pd.NA}
-    df = pd.DataFrame([{"input_identifier": ident, **meta[ident]} for ident in identifiers])
+    df = pd.DataFrame(
+        [{"input_identifier": ident, **meta[ident]} for ident in identifiers]
+    )
 
     # --- Early return for empty inputs ---
     if df.empty:
@@ -200,8 +208,14 @@ def get_properties(
     standard_props_snake = [p for p in resolved_props if p not in _SPECIAL_PROPS]
     if standard_props_snake:
         # Define internal fallback mappings (these are not exposed as separate properties)
-        fallback_map_snake = {"canonical_smiles": "connectivity_smiles", "isomeric_smiles": "fallback_smiles"}
-        internal_fallback_to_camel = {"connectivity_smiles": "ConnectivitySMILES", "fallback_smiles": "SMILES"}
+        fallback_map_snake = {
+            "canonical_smiles": "connectivity_smiles",
+            "isomeric_smiles": "fallback_smiles",
[…]
