### 2025-09-22 92b64a0
Fix CI and address code quality issues
