+++ .github/workflows/tests.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install ".[dev,plot]" pytest-cov coverage-badge
+          pip install ".[dev,plot]" pytest-cov coverage-badge mypy
 
       - name: Verify plotting libraries
         run: |
