### 2025-09-22 3493b43
Fix CI mypy and enhance docs syntax highlighting
