+++ .github/workflows/tests.yml
@@ -25,7 +25,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install ".[dev,plot]" pytest-cov coverage-badge mypy
+          pip install ".[dev,plot]" pytest-cov coverage-badge mypy types-requests
 
       - name: Verify plotting libraries
         run: |
@@ -43,7 +43,7 @@ jobs:
           ruff format --check src/ tests/
 
       - name: Run type checking with mypy
-        run: mypy src/ChemInformant --ignore-missing-imports
+        run: mypy src/ChemInformant --ignore-missing-imports --allow-untyped-defs --allow-incomplete-defs --no-strict-optional
 
       - name: Run tests with coverage
         env:
