### 2025-09-22 fe47c33
Fix mypy type annotations
