+++ src/ChemInformant/cheminfo_api.py
@@ -61,7 +61,7 @@ def get_properties(
     namespace: str = "cid",
     include_3d: bool = False,
     all_properties: bool = False,
-    **kwargs,
+    **kwargs: Any,
 ) -> pd.DataFrame:
     """
     Retrieve chemical properties for one or more compounds from PubChem.
@@ -185,19 +185,22 @@ def get_properties(
     if properties is not None and len(resolved_props) == 0:
         return pd.DataFrame()
 
+    identifiers_list: list[int | str]
     if not isinstance(identifiers, list):
-        identifiers = [identifiers]
+        identifiers_list = [identifiers]
+    else:
+        identifiers_list = list(identifiers)  # Type-safe conversion
 
     # --- Step 2: Create base DataFrame with resolved CIDs ---
     meta: dict[Any, dict[str, Any]] = {}
-    for ident in identifiers:
+    for ident in identifiers_list:
         try:
             cid = _resolve_to_single_cid(ident)
             meta[ident] = {"status": "OK", "cid": str(cid)}
         except Exception as exc:
             meta[ident] = {"status": type(exc).__name__, "cid": pd.NA}
     df = pd.DataFrame(
-        [{"input_identifier": ident, **meta[ident]} for ident in identifiers]
+        [{"input_identifier": ident, **meta[ident]} for ident in identifiers_list]
     )
 
     # --- Early return for empty inputs ---
@@ -232,12 +235,13 @@ def get_properties(
 
         for prop_snake in standard_props_snake:
             prop_camel = SNAKE_TO_CAMEL[prop_snake]
-            values = []
+            values: list[Any] = []
             for _, row in df.iterrows():
-                cid = int(row["cid"]) if pd.notna(row["cid"]) else None
+                cid_val = row["cid"]
+                cid = int(cid_val) if pd.notna(cid_val) else None
                 val = None
                 if row["status"] == "OK" and cid:
-                    api_row = fetched_data.get(cid, {})
+                    api_row = fetched_data.get(cid, {})  # type: ignore[unreachable]
                     val = api_row.get(prop_camel)
                     if not val and prop_snake in fallback_map_snake:
                         fallback_prop = fallback_map_snake[prop_snake]
@@ -255,7 +259,7 @@ def get_properties(
                 if prop_snake == "cas"
                 else api_helpers.get_synonyms_for_cid
             )
-            fail_value = None if prop_snake == "cas" else []
+            fail_value: Any = None if prop_snake == "cas" else []
             prop_data = [
                 fetch_func(int(cid)) if pd.notna(cid) and status == "OK" else fail_value
                 for cid, status in zip(df["cid"], df["status"])
@@ -281,7 +285,7 @@ def get_properties(
 # --- Convenience Functions (now simple and consistent) ---
 
 
-def _fetch_scalar(id_, prop_snake):
+def _fetch_scalar(id_: str | int, prop_snake: str) -> float | int | str | None:
     """Internal helper for single-value convenience functions."""
     try:
         cid = _resolve_to_single_cid(id_)
@@ -312,9 +316,10 @@ def _fetch_scalar(id_, prop_snake):
         props = api_helpers.get_batch_properties([cid], props_to_fetch)
         data = props.get(cid, {})
 
-        return data.get(prop_camel) or (
+        result = data.get(prop_camel) or (
[…]
