### 2025-09-22 6f0d19b
Fix ruff formatting for api_helpers.py
