### 2025-09-22 c841d15
Fix remaining mypy type annotation errors
