### 2025-09-22 5a96707
Remove unused type ignore comment
