### 2025-09-22 e33c787
Remove type ignore comment completely - CI doesn't need it
