### 2025-09-22 2610ef2
Update coverage badge [skip ci] 4e38a806aa00f1a5be6bd955f0f164bd7fb11b15
