+++ src/ChemInformant/constants.py
@@ -2,37 +2,68 @@
 Centralised property lists and aliases for the get_properties() function.
 This file establishes snake_case as the canonical internal format.
 """
+
 import re
 
 # --- 1. Master List of all Official PubChem Property Tags (in CamelCase) ---
 # This list represents the exact tags the PubChem API uses.
 ALL_PROPS_CAMEL: list[str] = [
-    "MolecularFormula", "MolecularWeight", "CanonicalSMILES", "IsomericSMILES",
-    "InChI", "InChIKey", "IUPACName", "XLogP", "ExactMass", "MonoisotopicMass",
-    "TPSA", "Complexity", "Charge", "HBondDonorCount", "HBondAcceptorCount",
-    "RotatableBondCount", "HeavyAtomCount", "IsotopeAtomCount", "AtomStereoCount",
-    "DefinedAtomStereoCount", "UndefinedAtomStereoCount", "BondStereoCount",
-    "DefinedBondStereoCount", "UndefinedBondStereoCount", "CovalentUnitCount",
-    "Volume3D", "XStericQuadrupole3D", "YStericQuadrupole3D",
-    "ZStericQuadrupole3D", "FeatureCount3D", "FeatureAcceptorCount3D",
-    "FeatureDonorCount3D", "FeatureAnionCount3D", "FeatureCationCount3D",
-    "FeatureRingCount3D", "FeatureHydrophobeCount3D", "ConformerModelRMSD3D",
-    "EffectiveRotorCount3D", "ConformerCount3D"
+    "MolecularFormula",
+    "MolecularWeight",
+    "CanonicalSMILES",
+    "IsomericSMILES",
+    "InChI",
+    "InChIKey",
+    "IUPACName",
+    "XLogP",
+    "ExactMass",
+    "MonoisotopicMass",
+    "TPSA",
+    "Complexity",
+    "Charge",
+    "HBondDonorCount",
+    "HBondAcceptorCount",
+    "RotatableBondCount",
+    "HeavyAtomCount",
+    "IsotopeAtomCount",
+    "AtomStereoCount",
+    "DefinedAtomStereoCount",
+    "UndefinedAtomStereoCount",
+    "BondStereoCount",
+    "DefinedBondStereoCount",
+    "UndefinedBondStereoCount",
+    "CovalentUnitCount",
+    "Volume3D",
+    "XStericQuadrupole3D",
+    "YStericQuadrupole3D",
+    "ZStericQuadrupole3D",
+    "FeatureCount3D",
+    "FeatureAcceptorCount3D",
+    "FeatureDonorCount3D",
+    "FeatureAnionCount3D",
+    "FeatureCationCount3D",
+    "FeatureRingCount3D",
+    "FeatureHydrophobeCount3D",
+    "ConformerModelRMSD3D",
+    "EffectiveRotorCount3D",
+    "ConformerCount3D",
 ]
+
 
 def _snake_case(s: str) -> str:
     """Converts CamelCase to snake_case, handling specifics like '3D'."""
     # First handle 3D suffix specifically
-    s = s.replace('3D', '_3D')
+    s = s.replace("3D", "_3D")
 
     # Standard CamelCase to snake_case conversion
-    s1 = re.sub('([a-z0-9])([A-Z])', r'\1_\2', s)
-    s2 = re.sub('([A-Z])([A-Z][a-z])', r'\1_\2', s1)
+    s1 = re.sub("([a-z0-9])([A-Z])", r"\1_\2", s)
+    s2 = re.sub("([A-Z])([A-Z][a-z])", r"\1_\2", s1)
 
     # Convert to lowercase and clean up any double underscores
-    result = s2.lower().replace('__', '_')
+    result = s2.lower().replace("__", "_")
 
     return result
+
 
[…]
