+++ src/ChemInformant/cheminfo_api.py
@@ -2,12 +2,14 @@
 This module contains the high-level, user-facing API for ChemInformant.
 It now operates exclusively on a snake_case standard for all property names.
 """
+
 from __future__ import annotations
 
 import logging
 import re
 import sys
-from typing import Any, Iterable
+from collections.abc import Iterable
+from typing import Any
 
 import pandas as pd
 
@@ -24,8 +26,10 @@ from .models import AmbiguousIdentifierError, Compound, NotFoundError
 _SMILES_TOKENS = re.compile(r"[=#\[\]\(\)]|\d|Br|Cl|Si", re.I)
 _SPECIAL_PROPS = {"cas", "synonyms"}
 
+
 def _looks_like_smiles(s: str) -> bool:
     return bool(_SMILES_TOKENS.search(s))
+
 
 def _resolve_to_single_cid(identifier: str | int) -> int:
     # This function remains robust as is.
@@ -49,6 +53,7 @@ def _resolve_to_single_cid(identifier: str | int) -> int:
             raise AmbiguousIdentifierError(identifier, cids)
     raise NotFoundError(identifier)
 
+
 def get_properties(
     identifiers: int | str | list[int] | list[str],
     properties: str | list[str] | None = None,
@@ -56,7 +61,7 @@ def get_properties(
     namespace: str = "cid",
     include_3d: bool = False,
     all_properties: bool = False,
-    **kwargs,
+    **kwargs: Any,
 ) -> pd.DataFrame:
     """
     Retrieve chemical properties for one or more compounds from PubChem.
@@ -143,7 +148,9 @@ def get_properties(
     """
     # --- Parameter validation ---
     if all_properties and (properties is not None or include_3d):
-        raise ValueError("`all_properties=True` is mutually exclusive with `properties` and `include_3d`.")
+        raise ValueError(
+            "`all_properties=True` is mutually exclusive with `properties` and `include_3d`."
+        )
     if properties is not None and include_3d:
         logging.warning("When `properties` is specified, `include_3d` is ignored.")
 
@@ -155,7 +162,7 @@ def get_properties(
         original_props_order = resolved_props
     elif properties is not None:
         if isinstance(properties, str):
-            properties = [p.strip() for p in properties.split(',')]
+            properties = [p.strip() for p in properties.split(",")]
         original_props_order = list(properties)
         resolved_props = []
         unsupported_props = []
@@ -168,7 +175,7 @@ def get_properties(
 
         if unsupported_props:
             raise ValueError(f"Unsupported properties: {unsupported_props}")
-    else: # Default behavior when properties is None
+    else:  # Default behavior when properties is None
         resolved_props = CORE_PROPS[:]
         if include_3d:
             resolved_props.extend(THREED_PROPS)
@@ -178,18 +185,23 @@ def get_properties(
     if properties is not None and len(resolved_props) == 0:
         return pd.DataFrame()
 
+    identifiers_list: list[int | str]
     if not isinstance(identifiers, list):
-        identifiers = [identifiers]
+        identifiers_list = [identifiers]
+    else:
+        identifiers_list = list(identifiers)  # Type-safe conversion
 
     # --- Step 2: Create base DataFrame with resolved CIDs ---
     meta: dict[Any, dict[str, Any]] = {}
-    for ident in identifiers:
+    for ident in identifiers_list:
         try:
             cid = _resolve_to_single_cid(ident)
             meta[ident] = {"status": "OK", "cid": str(cid)}
[…]
