### 2025-09-22 9bb4e8a
Further improve docs syntax highlighting consistency
