### 2025-09-22 1de0d7b
Update coverage badge [skip ci] 0086e6595d94e55ddb89d11f53e6728cfa2101ab
