+++ docs/requirements.txt
@@ -1,6 +1,6 @@
-sphinx
-sphinx-rtd-theme
-sphinx-copybutton
+sphinx>=5.0.0
+sphinx-rtd-theme>=1.2.0
+sphinx-copybutton>=0.5.0
 
-# Required for consistent syntax highlighting
+# Required for consistent and enhanced syntax highlighting
 pygments>=2.17.0
