### 2025-09-22 #254 @HzaCode
Hi @eliotwrobson,

Thank you for the quick and very helpful review! I also really appreciate the work you and the other editors have done to establish the JOSS -> pyOS fast-track process, and I'm honored to be a part of it.

I've opened a few pull requests to address the points you raised.

Specifically, [PR #36](https://github.com/HzaCode/ChemInformant/pull/36) adds the docs badge, updates the README, and moves the benchmarks. It also adds a `ruff` workflow and drops support for older Python versions. The `mypy` typing issues are fixed in [PR #37](https://github.com/HzaCode/ChemInformant/pull/37), and the documentation syntax highlighting is corrected in [PR #38](https://github.com/HzaCode/ChemInformant/pull/38).

Thanks again for your guidance. Please let me know if anything else is needed.
