### 2025-09-25 cbd8c63
Bump version to 2.4.3 and update documentation syntax highlighting
