### 2025-09-29 #254 @HzaCode
@eliotwrobson Thanks for your response! I'm glad to hear that everything important has been handled. I'm looking forward to seeing my package on the pyOpenSci "Accepted packages" list.
