### 2025-09-30 #254 @eliotwrobson
----------------------------------------------
🎉 ChemInformant has been approved by pyOpenSci! Thank you @HzaCode  for submitting ChemInformant and many thanks to the prior JOSS reviewers and editor for their work on this package! 😸

## Author Wrap Up Tasks

- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/254)`.
- [x] Add ChemInformant to the pyOpenSci website. @HzaCode, please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and name to the list of contributors
- [x] @HzaCode You are welcome to write a blog post if you are interested - here are a few examples of other blog posts: [pandera](https://www.pyopensci.org/blog/pandera-python-pandas-dataframe-validation) [movingpandas](https://www.pyopensci.org/blog/movingpandas-movement-data-analysis), and [here](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) is a markdown example that you could use as a guide when creating your post. Here is a [prior PR for a blog post](https://github.com/pyOpenSci/pyopensci.github.io/pull/433). This can even be a tutorial on how to use the package. No worries if not though, but we'd love to spread the word about your package!
- [x] @HzaCode will send you a slack invite on the email on your GitHub as soon as I can. Let me know if you don't get the invite after a couple days.

----------------------------------------------

If you have any feedback for us about the review process please feel free to share it here. We are always looking to improve our process and our documentation in the [contributing-guide](https://www.pyopensci.org/contributing-guide). We have also been updating our documentation to improve the process so all feedback is appreciated!
