### 2025-09-30 dee5cf9
Update CITATION.cff version to 2.4.3 for Zenodo backup
