### 2025-09-30 63fe1f3
Update coverage badge [skip ci] d5b7dfbb46f5afd63af81ab50e413fb24c4f1827
