+++ conda-recipe/meta.yaml
@@ -1,0 +1,56 @@
+{% set name = "cheminformant" %}
+{% set version = "2.4.2" %}
+
+package:
+  name: {{ name|lower }}
+  version: {{ version }}
+
+source:
+  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/cheminformant-{{ version }}.tar.gz
+  sha256: # 这里需要填入PyPI上tar.gz文件的SHA256哈希值
+
+build:
+  noarch: python
+  script: {{ PYTHON }} -m pip install . -vv
+  number: 0
+
+requirements:
+  host:
+    - python >=3.8
+    - pip
+    - setuptools
+  run:
+    - python >=3.8
+    - requests >=2.25.0
+    - requests-cache >=1.2.0
+    - pydantic >=2.0
+    - pandas >=1.5.0
+    - sqlalchemy >=1.4
+    - pystow >=0.7
+
+test:
+  imports:
+    - ChemInformant
+  commands:
+    - chemfetch --help
+    - chemdraw --help
+  requires:
+    - pip
+
+about:
+  home: https://github.com/HzaCode/ChemInformant
+  summary: A robust and high-throughput Python client for the PubChem API, designed for automated data retrieval and analysis
+  description: |
+    ChemInformant is a robust data acquisition engine for the PubChem database, 
+    engineered for the modern scientific workflow. It intelligently manages network 
+    requests, performs rigorous runtime data validation, and delivers analysis-ready 
+    results, providing a dependable foundation for any computational chemistry project in Python.
+  license: MIT
+  license_file: LICENSE
+  doc_url: https://hezhiang.com/ChemInformant
+  dev_url: https://github.com/HzaCode/ChemInformant
+
+extra:
+  recipe-maintainers:
+    - HzaCode  # 你的GitHub用户名
+
