### 2025-09-30 #254 @HzaCode
Thank you very much to you and the entire pyOpenSci team for the approval of ChemInformant! I also want to extend my gratitude to the previous JOSS reviewers and editor for their valuable contributions.

I have now completed all the wrap-up tasks for version 2.4.3:

*   ✅ Activated Zenodo to watch the repository.
*   ✅ Tagged and created the 2.4.3 release to generate a Zenodo version and DOI.
*   ✅ Added the pyOpenSci peer-review badge to the `README.md`.
*   ✅ Filled out the post-review survey.
*   ✅ Opened a pull request to add ChemInformant to the pyOpenSci website.

I appreciate the invitation to write a blog post and I am interested in doing so. I will work on that and submit it later on.

I will look out for the Slack invitation. Thanks again for all your help and support
