### 2025-10-07 #254 @pllim
I think this is failing a workflow at https://github.com/pyOpenSci/pyopensci.github.io/actions/runs/18294498220/job/52168625783
